# backend/docs/swagger.yaml
basePath: /
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  models.AssignBackupValidatorRequest:
    properties:
      backupValidatorId:
        type: string
      reason:
        type: string
    type: object
...
  models.ResetValidatorDeviceRequest:
    properties:
      reason:
        type: string
    type: object
...
  models.ScanLog:
    properties:
      actorId:
        type: string
      aiClass:
        type: string
      aiConfidence:
        type: number
      batchId:
        type: string
      boxId:
        type: string
      createdAt:
        type: string
      durationMs:
        type: integer
      holdingTempC:
        type: number
      id:
        type: string
      imageRef:
        type: string
      qrToken:
        type: string
      reason:
        type: string
      releaseTempC:
        type: number
      verdict:
        type: string
      visualScore:
        type: number
    type: object
...
  models.SetValidatorStatusRequest:
    properties:
      reason:
        type: string
      status:
        type: string
    type: object
...
  models.ValidatorProfile:
    properties:
      anomalies:
        type: integer
      avgDurationSec:
        type: number
      backupValidatorId:
        type: string
      backupValidatorName:
        type: string
      certification:
        type: string
      device:
        type: string
      deviceBound:
        type: boolean
      deviceId:
        type: string
      email:
        type: string
      id:
        type: string
      lastScanAt:
        type: string
      lastWarningAt:
        type: string
      lastWarningNote:
        type: string
      name:
        type: string
      nip:
        type: string
      npsn:
        type: string
      quotaToday:
        type: integer
      role:
        type: string
      satgasId:
        type: string
      scanLogs:
        items:
          $ref: '#/definitions/models.ScanLog'
        type: array
      scansToday:
        type: integer
      schoolCity:
        type: string
      schoolName:
        type: string
      status:
        type: string
      userId:
        type: string
      warningCount:
        type: integer
    type: object
  models.WarnValidatorRequest:
    properties:
      note:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://localhost:8080
  description: RESTful API Service built with Golang, Clean Code, and SOLID principles.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: MBG Backend API
  version: "1.0"
paths:
...
  /validators/{id}/backup:
    put:
      consumes:
      - application/json
      description: Cadangan wajib validator aktif pada sekolah yang sama.
      parameters:
      - description: Validator ID
        example: VAL-001
        in: path
        name: id
        required: true
        type: string
      - description: ID cadangan + alasan
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.AssignBackupValidatorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ValidatorProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Tugaskan guru piket cadangan
      tags:
      - validators
  /validators/{id}/device/reset:
    post:
      consumes:
      - application/json
      description: Memutus tautan kriptografis perangkat; validator wajib registrasi
        ulang.
      parameters:
      - description: Validator ID
        example: VAL-001
        in: path
        name: id
        required: true
        type: string
      - description: Alasan reset
        in: body
        name: payload
        schema:
          $ref: '#/definitions/models.ResetValidatorDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ValidatorProfile'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Reset device binding validator
      tags:
      - validators
  /validators/{id}/status:
    patch:
      consumes:
      - application/json
      description: Superadmin saja. Menyelaraskan validator_profiles.status dengan
        users.status dan mencatat audit.
      parameters:
      - description: Validator ID
        example: VAL-001
        in: path
        name: id
        required: true
        type: string
      - description: Status baru + alasan
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.SetValidatorStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ValidatorProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Ubah status validator lapangan
      tags:
      - validators
  /validators/{id}/warnings:
    post:
      consumes:
      - application/json
      description: Menandai profil sebagai flagged dan menambah hitungan peringatan.
      parameters:
      - description: Validator ID
        example: VAL-001
        in: path
        name: id
        required: true
        type: string
      - description: Isi surat peringatan
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.WarnValidatorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ValidatorProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Kirim surat peringatan digital
      tags:
      - validators
swagger: "2.0"